blogapp/requirements.txt
absl-py==1.0.0
APScheduler==3.6.3
asgiref==3.5.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
backoff==1.11.1
beautifulsoup4==4.10.0
branca==0.4.2
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
comtypes==1.1.10
cryptography==36.0.1
cycler==0.11.0
daytime==0.4
distlib==0.3.4
django-admin-data-from-external-service==0.1.1
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
emoji==1.6.1
enum34==1.1.10
et-xmlfile==1.1.0
ExifRead==2.3.2
filelock==3.4.2
folium==0.12.1.post1
fonttools==4.28.3
freegames==2.3.2
future==0.18.2
googletrans==3.0.0
gunicorn==20.1.0
gym==0.21.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.13.5
importlib-metadata==4.11.0
iniconfig==1.1.1
Jinja2==3.0.3
joblib==1.1.0
keyboard==0.13.5
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
mediapipe==0.8.9.1
MouseInfo==0.1.3
mysql==0.0.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
networkx==2.6.3
nltk==3.6.5
numpy==1.21.4
opencage==2.0.0
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
packbits==0.6
Panda3D==1.10.10
panda3d-gltf==0.13
panda3d-simplepbr==0.9
pandas==1.3.4
phonenumbers==8.12.38
Pillow==8.4.0
Pint==0.18
pip-install-test==0.5
platformdirs==2.4.1
pluggy==1.0.0
protobuf==3.19.1
psd-tools3==1.8.2
psutil==5.9.0
py==1.11.0
PyAutoGUI==0.9.53
PyBundle==1.0.6
pycaw==20181226
pycparser==2.21
pydantic==1.8.2
pygame==2.1.0
PyGetWindow==0.0.9
pyglet==1.5.11
pymatrix-rain==1.1.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.6
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pypiwin32==223
pyreadline3==3.3
PyRect==0.1.4
PyScreeze==0.1.28
pytesseract==0.3.8
pytest==6.2.5
python-dateutil==2.8.2
python-telegram-bot==13.10
python-utils==2.5.6
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pywin32==302
questions==0.7.0a4
regex==2021.11.10
requests==2.26.0
rfc3986==1.5.0
scikit-image==0.19.1
scipy==1.7.3
screeninfo==0.7
secure-smtplib==0.1.1
selenium==4.1.0
sequence==0.3.4
simpleeval==0.9.11
six==1.16.0
Snake==0.2
SnakeGame==1.0.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
speedtest-cli==2.1.3
sqlparse==0.4.2
tabulate==0.8.9
textblob==0.17.1
tifffile==2021.11.2
toml==0.10.2
tornado==6.1
tox==3.24.5
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
Unidecode==1.3.2
urllib3==1.26.7
# Editable install with no version control (ursina==4.1.1)
-e c:\users\user\appdata\local\programs\python\python310\lib\site-packages
vector==0.8.4
virtualenv==20.13.0
whitenoise==6.0.0
windows-curses==2.3.0
wsproto==1.0.0
zipp==3.7.0
